separated ShoppingProduct from TaskToDo
