+++ ToDoApplication/Models/TaskToDo.cs
@@ -33,15 +33,5 @@ namespace ToDoApplication.Models
         ToDo, InProgress, Done
     }
 
-    public class ShoppingProduct
-    {
-        [Key]
-        public int productId { get; set; }
-        public string name { get; set; }
-        public int quantity { get; set; }
 
-        [ForeignKey("TaskToDo")]
-        public int? TaskToDoId { get; set; }
-        public TaskToDo? TaskToDo { get; set; }
-    }
 }
