+++ ToDoApplication/Services/Interfaces/IShoppingProductService.cs
@@ -1,0 +1,9 @@
+﻿using ToDoApplication.Models;
+
+namespace ToDoApplication.Services.Interfaces
+{
+    public interface IShoppingProductService
+    {
+        int Edit(int shoppingProductId, ShoppingProduct body);
+    }
+}
